refactor(codex): isolate native MCP App contracts
